New minimap; skill, trap, map, collision and AI changes; new animations, etc.

+++ Assets/Units/ISkill.cs
@@ -1,11 +1,10 @@
 using UnityEngine;
 
-interface ISkill{
-	
-	Texture2D getSkillIcon();
-	float getCurrentCooldown();
-	int getMaxCooldown();
-	float getMaxRange();
-
+interface ISkill
+{
+    Texture2D getSkillIcon();
+    float getCurrentCooldown();
+    float getMaxCooldown();
+    float getMaxRange();
 }
